+++ modules/core.lua
@@ -2,5 +2,5 @@
 EZOArmory = EZOArmory or {}
 
 EZOArmory.ADDON_NAME    = "EZOArmory"
-EZOArmory.ADDON_VERSION = "0.8.5"
+EZOArmory.ADDON_VERSION = "0.8.6"
 EZOArmory.AUTHOR        = "@Zuriplayer"